+++ provision/provision.sh
@@ -9,13 +9,16 @@
 # end of this script.
 start_seconds=`date +%s`
 
-# TODO: Grab this from the Vagrantfile
-vagrant_ip=192.168.50.4
-
-# Capture a basic ping result to one of Google's DNS servers to try and
-# determine if outside access is available to us. If it isn't, we'll
-# want to skip a few things in the future rather than creating a bunch of errors.
-ping_result=`ping -c 2 8.8.8.8 2>&1`
+# Capture a basic ping result to Google's primary DNS server to determine if
+# outside access is available to us. If this does not reply after 2 attempts,
+# we try one of Level3's DNS servers as well. If neither of these IPs replies to
+# a ping, then we'll skip a few things further in provisioning rather than
+# creating a bunch of errors.
+ping_result=`ping -c 2 8.8.4.4 2>&1`
+if [[ $ping_result != *bytes?from* ]]
+then
+	ping_result=`ping -c 2 4.2.2.2 2>&1`
+fi
 
 # PACKAGE INSTALLATION
 #
@@ -61,6 +64,9 @@ apt_package_check_list=(
 	# memcached is made available for object caching
 	memcached
 
+	# mysql is the default database
+	mysql-server
+
 	# other packages that come in handy
 	imagemagick
 	subversion
@@ -70,45 +76,52 @@ apt_package_check_list=(
 	curl
 	make
 	vim
+	colordiff
+
+	# Req'd for i18n tools
+	gettext
+
+	# Req'd for Webgrind
+	graphviz
 
 	# dos2unix
 	# Allows conversion of DOS style line endings to something we'll have less
 	# trouble with in Linux.
 	dos2unix
+
+	# nodejs for use by grunt
+	g++
+	nodejs
+
 )
 
-echo "Check for packages to install..."
+echo "Check for apt packages to install..."
 
 # Loop through each of our packages that should be installed on the system. If
 # not yet installed, it should be added to the array of packages to install.
 for pkg in "${apt_package_check_list[@]}"
 do
-	if dpkg -s $pkg 2>&1 | grep -q 'Status: install ok installed';
-	then 
-		echo "  * " $pkg already installed
+	package_version=`dpkg -s $pkg 2>&1 | grep 'Version:' | cut -d " " -f 2`
+	if [[ $package_version != "" ]]
+	then
+		space_count=`expr 20 - "${#pkg}"` #11
+		pack_space_count=`expr 30 - "${#package_version}"`
+		real_space=`expr ${space_count} + ${pack_space_count} + ${#package_version}`
+		printf " * $pkg %${real_space}.${#package_version}s ${package_version}\n"
 	else
-		echo "  * " $pkg not yet installed
+		echo " *" $pkg [not installed]
 		apt_package_install_list+=($pkg)
 	fi
 done
 
 # MySQL
 #
-# The current state of MySQL should be done outside of the looping done above.
-# This allows us to set the MySQL specific settings for the root password
-# so that provisioning does not require any user input.
-if dpkg -s mysql-server | grep -q 'Status: install ok installed';
-then
-	echo "  * mysql-server already installed"
-else 
-	echo "  * mysql-server not yet installed"
-	# We need to set the selections to automatically fill the password prompt
-	# for mysql while it is being installed. The password in the following two
-	# lines *is* actually set to the word 'blank' for the root user.
-	echo mysql-server mysql-server/root_password password blank | debconf-set-selections
-	echo mysql-server mysql-server/root_password_again password blank | debconf-set-selections
-	apt_package_install_list+=('mysql-server')
-fi
+# Use debconf-set-selections to specify the default password for the root MySQL
+# account. This runs on every provision, even if MySQL has been installed. If
+# MySQL is already installed, it will not affect anything. The password in the
+# following two lines *is* actually set to the word 'blank' for the root user.
+echo mysql-server mysql-server/root_password password blank | debconf-set-selections
+echo mysql-server mysql-server/root_password_again password blank | debconf-set-selections
 
 # Provide our custom apt sources before running `apt-get update`
 ln -sf /srv/config/apt-source-append.list /etc/apt/sources.list.d/vvv-sources.list | echo "Linked custom apt sources"
@@ -118,8 +131,8 @@ then
 	# If there are any packages to be installed in the apt_package_list array,
 	# then we'll run `apt-get update` and then `apt-get install` to proceed.
 	if [ ${#apt_package_install_list[@]} = 0 ];
-	then 
-		printf "No packages to install.\n\n"
+	then
+		echo -e "No apt packages to install.\n"
 	else
 		# Before running `apt-get update`, we should add the public keys for
 		# the packages that we are installing from non standard sources via
@@ -141,12 +154,16 @@ then
 		gpg -q --keyserver keyserver.ubuntu.com --recv-key A1715D88E1DF1F24
 		gpg -q -a --export A1715D88E1DF1F24 | apt-key add -
 
+		# Launchpad nodejs key C7917B12
+		gpg -q --keyserver keyserver.ubuntu.com --recv-key C7917B12
+		gpg -q -a --export  C7917B12  | apt-key add -
+
 		# update all of the package references before installing anything
-		printf "Running apt-get update....\n"
+		echo "Running apt-get update..."
 		apt-get update --assume-yes
 
 		# install required packages
-		printf "Installing apt-get packages...\n"
+		echo "Installing apt-get packages..."
 		apt-get install --assume-yes ${apt_package_install_list[@]}
 
 		# Clean up apt caches
@@ -171,10 +188,10 @@ then
 	# master branch on GitHub repository.
 	if composer --version | grep -q 'Composer version';
 	then
-		printf "Updating Composer...\n"
+		echo "Updating Composer..."
 		composer self-update
 	else
-		printf "Installing Composer...\n"
+		echo "Installing Composer..."
 		curl -sS https://getcomposer.org/installer | php
 		chmod +x composer.phar
 		mv composer.phar /usr/local/bin/composer
@@ -187,34 +204,63 @@ then
 	# packages are controlled in the `/srv/config/phpunit-composer.json` file.
 	if [ ! -d /usr/local/src/vvv-phpunit ]
 	then
-		printf "Installing PHPUnit, Hamcrest and Mockery...\n"
+		echo "Installing PHPUnit, Hamcrest and Mockery..."
 		mkdir -p /usr/local/src/vvv-phpunit
 		cp /srv/config/phpunit-composer.json /usr/local/src/vvv-phpunit/composer.json
 		sh -c "cd /usr/local/src/vvv-phpunit && composer install"
 	else
 		cd /usr/local/src/vvv-phpunit
-		if composer show -i | grep -q 'mockery'; then echo 'Mockery installed';else vvvphpunit_update=1;fi
-		if composer show -i | grep -q 'phpunit'; then echo 'PHPUnit installed'; else vvvphpunit_update=1;fi
-		if composer show -i | grep -q 'hamcrest'; then echo 'Hamcrest installed'; else vvvphpunit_update=1;fi
+		if composer show -i | grep -q 'mockery' ; then echo "Mockery installed" ; else vvvphpunit_update=1; fi
+		if composer show -i | grep -q 'phpunit' ; then echo "PHPUnit installed" ; else vvvphpunit_update=1; fi
+		if composer show -i | grep -q 'hamcrest'; then echo "Hamcrest installed"; else vvvphpunit_update=1; fi
 		cd ~/
 	fi
 
 	if [ "$vvvphpunit_update" = 1 ]
 	then
-		printf "Update PHPUnit, Hamcrest and Mockery...\n"
+		echo "Update PHPUnit, Hamcrest and Mockery..."
 		cp /srv/config/phpunit-composer.json /usr/local/src/vvv-phpunit/composer.json
 		sh -c "cd /usr/local/src/vvv-phpunit && composer update"
 	fi
+
+	# Grunt
+	#
+	# Install or Update Grunt based on gurrent state.  Updates are direct
+	# from NPM
+	if grunt --version ;
+	then
+		echo "Updating Grunt CLI"
+		npm update -g grunt-cli
+	else
+		echo "Installing Grunt CLI"
+		npm install -g grunt-cli
+	fi
+
 else
-	printf "\nNo network connection available, skipping package installation"
+	echo -e "\nNo network connection available, skipping package installation"
+fi
+
+# Configuration for nginx
+if [ ! -e /etc/nginx/server.key ]; then
+	echo "Generate Nginx server private key..."
+	vvvgenrsa=`openssl genrsa -out /etc/nginx/server.key 2048 2>&1`
+	echo $vvvgenrsa
+fi
+if [ ! -e /etc/nginx/server.csr ]; then
+	echo "Generate Certificate Signing Request (CSR)..."
+	openssl req -new -batch -key /etc/nginx/server.key -out /etc/nginx/server.csr
+fi
+if [ ! -e /etc/nginx/server.crt ]; then
+	echo "Sign the certificate using the above private key and CSR..."
+	vvvsigncert=`openssl x509 -req -days 365 -in /etc/nginx/server.csr -signkey /etc/nginx/server.key -out /etc/nginx/server.crt 2>&1`
+	echo $vvvsigncert
 fi
 
 # SYMLINK HOST FILES
-printf "\nLink Directories...\n"
+echo -e "\nSetup configuration file links..."
 
-# Configuration for nginx
-ln -sf /srv/config/nginx-config/nginx.conf /etc/nginx/nginx.conf | echo "Linked nginx.conf to /etc/nginx/"
-ln -sf /srv/config/nginx-config/nginx-wp-common.conf /etc/nginx/nginx-wp-common.conf | echo "Linked nginx-wp-common.conf to /etc/nginx/"
+ln -sf /srv/config/nginx-config/nginx.conf /etc/nginx/nginx.conf | echo " * /srv/config/nginx-config/nginx.conf -> /etc/nginx/nginx.conf"
+ln -sf /srv/config/nginx-config/nginx-wp-common.conf /etc/nginx/nginx-wp-common.conf | echo " * /srv/config/nginx-config/nginx-wp-common.conf -> /etc/nginx/nginx-wp-common.conf"
 
 # /etc/nginx/symlinked-confs
 symlink_conf_dir=/etc/nginx/symlinked-confs
@@ -229,40 +275,49 @@ for site_config_file in $(find /srv/www -name $symlink_target_filename); do
 done
 
 # Configuration for php5-fpm
-ln -sf /srv/config/php5-fpm-config/www.conf /etc/php5/fpm/pool.d/www.conf | echo "Linked www.conf to /etc/php5/fpm/pool.d/"
+ln -sf /srv/config/php5-fpm-config/www.conf /etc/php5/fpm/pool.d/www.conf | echo " * /srv/config/php5-fpm-config/www.conf -> /etc/php5/fpm/pool.d/www.conf"
 
 # Provide additional directives for PHP in a custom ini file
-ln -sf /srv/config/php5-fpm-config/php-custom.ini /etc/php5/fpm/conf.d/php-custom.ini | echo "Linked php-custom.ini to /etc/php5/fpm/conf.d/php-custom.ini"
+ln -sf /srv/config/php5-fpm-config/php-custom.ini /etc/php5/fpm/conf.d/php-custom.ini | echo " * /srv/config/php5-fpm-config/php-custom.ini -> /etc/php5/fpm/conf.d/php-custom.ini"
 
-# Configuration for Xdebug - Mod disabled by default
-php5dismod xdebug
-ln -sf /srv/config/php5-fpm-config/xdebug.ini /etc/php5/fpm/conf.d/xdebug.ini | echo "Linked xdebug.ini to /etc/php5/fpm/conf.d/xdebug.ini"
+# Configuration for Xdebug
+ln -sf /srv/config/php5-fpm-config/xdebug.ini /etc/php5/fpm/conf.d/xdebug.ini | echo " * /srv/config/php5-fpm-config/xdebug.ini -> /etc/php5/fpm/conf.d/xdebug.ini"
 
 # Configuration for APC
-ln -sf /srv/config/php5-fpm-config/apc.ini /etc/php5/fpm/conf.d/apc.ini | echo "Linked apc.ini to /etc/php5/fpm/conf.d/"
+ln -sf /srv/config/php5-fpm-config/apc.ini /etc/php5/fpm/conf.d/apc.ini | echo " * /srv/config/php5-fpm-config/apc.ini -> /etc/php5/fpm/conf.d/apc.ini"
 
 # Configuration for mysql
-cp /srv/config/mysql-config/my.cnf /etc/mysql/my.cnf | echo "Linked my.cnf to /etc/mysql/"
+cp /srv/config/mysql-config/my.cnf /etc/mysql/my.cnf | echo " * /srv/config/mysql-config/my.cnf -> /etc/mysql/my.cnf"
 
 # Configuration for memcached
-ln -sf /srv/config/memcached-config/memcached.conf /etc/memcached.conf | echo "Linked memcached.conf to /etc/"
+ln -sf /srv/config/memcached-config/memcached.conf /etc/memcached.conf | echo " * /srv/config/memcached-config/memcached.conf -> /etc/memcached.conf"
 
 # Custom bash_profile for our vagrant user
-ln -sf /srv/config/bash_profile /home/vagrant/.bash_profile | echo "Linked .bash_profile to vagrant user's home directory..."
+ln -sf /srv/config/bash_profile /home/vagrant/.bash_profile | echo " * /srv/config/bash_profile -> /home/vagrant/.bash_profile"
 
 # Custom bash_aliases included by vagrant user's .bashrc
-ln -sf /srv/config/bash_aliases /home/vagrant/.bash_aliases | echo "Linked .bash_aliases to vagrant user's home directory..."
+ln -sf /srv/config/bash_aliases /home/vagrant/.bash_aliases | echo " * /srv/config/bash_aleases -> /home/vagrant/.bash_aliases"
+
+# Custom home bin directory
+ln -nsf /srv/config/homebin /home/vagrant/bin | echo " * /srv/config/homebin -> /home/vagrant/bin"
 
 # Custom vim configuration via .vimrc
-ln -sf /srv/config/vimrc /home/vagrant/.vimrc | echo "Linked vim configuration to home directory..."
+ln -sf /srv/config/vimrc /home/vagrant/.vimrc | echo " * /srv/config/vimrc -> /home/vagrant/.vimrc"
+
+# Capture the current IP address of the virtual machine into a variable that
+# can be used when necessary throughout provisioning.
+vvv_ip=`ifconfig eth1 | ack "inet addr" | cut -d ":" -f 2 | cut -d " " -f 1`
 
 # RESTART SERVICES
 #
 # Make sure the services we expect to be running are running.
-printf "\nRestart services...\n"
+echo -e "\nRestart services..."
 service nginx restart
-service php5-fpm restart
 service memcached restart
+
+# Disable PHP Xdebug module by default
+php5dismod xdebug
+service php5-fpm restart
 
 # MySQL gives us an error if we restart a non running service, which
 # happens after a `vagrant halt`. Check to see if it's running before
@@ -270,10 +325,10 @@ service memcached restart
 exists_mysql=`service mysql status`
 if [ "mysql stop/waiting" == "$exists_mysql" ]
 then
-	printf "service mysql start"
+	echo "service mysql start"
 	service mysql start
 else
-	printf "service mysql restart"
+	echo "service mysql restart"
 	service mysql restart
 fi
 
@@ -283,9 +338,9 @@ fi
 # the mysqldump files located in database/backups/
 if [ -f /srv/database/init-custom.sql ]
 then
-	mysql -u root -pblank < /srv/database/init-custom.sql | printf "\nInitial custom MySQL scripting...\n"
+	mysql -u root -pblank < /srv/database/init-custom.sql | echo -e "\nInitial custom MySQL scripting..."
 else
-	printf "\nNo custom MySQL scripting found in database/init-custom.sql, skipping...\n"
+	echo -e "\nNo custom MySQL scripting found in database/init-custom.sql, skipping..."
 fi
 
 for init_sql_file in $(find /srv/www/ -name vvv-init.sql)
@@ -296,21 +351,22 @@ done
 
 # Setup MySQL by importing an init file that creates necessary
 # users and databases that our vagrant setup relies on.
-mysql -u root -pblank < /srv/database/init.sql | echo "Initial MySQL prep...."
+mysql -u root -pblank < /srv/database/init.sql | echo "Initial MySQL prep..."
 
 # WP-CLI Install before import-sql.sh since it depends on it
 if [[ $ping_result == *bytes?from* ]]
 then
 	if [ ! -d /srv/www/wp-cli ]
 	then
-		printf "\nDownloading wp-cli.....http://wp-cli.org\n"
+		echo -e "\nDownloading wp-cli, see http://wp-cli.org"
 		git clone git://github.com/wp-cli/wp-cli.git /srv/www/wp-cli
 		cd /srv/www/wp-cli
 		composer install
 	else
-		printf "\nUpdating wp-cli....\n"
+		echo -e "\nUpdating wp-cli..."
 		cd /srv/www/wp-cli
 		git pull --rebase origin master
+		composer update
 	fi
 	# Link `wp` to the `/usr/local/bin` directory
 	ln -sf /srv/www/wp-cli/bin/wp /usr/local/bin/wp
@@ -322,85 +378,129 @@ fi
 
 if [[ $ping_result == *bytes?from* ]]
 then
+	# Download and extract phpMemcachedAdmin to provide a dashboard view and admin interface
+	# to the goings on of memcached when running
+	if [ ! -d /srv/www/default/memcached-admin ]
+	then
+		echo -e "\nDownloading phpMemcachedAdmin, see https://code.google.com/p/phpmemcacheadmin/"
+		cd /srv/www/default
+		wget -q -O phpmemcachedadmin.tar.gz 'https://phpmemcacheadmin.googlecode.com/files/phpMemcachedAdmin-1.2.2-r262.tar.gz'
+		mkdir memcached-admin
+		tar -xf phpmemcachedadmin.tar.gz --directory memcached-admin
+		rm phpmemcachedadmin.tar.gz
+	else
+		echo "phpMemcachedAdmin already installed."
+	fi
+
+	# Webgrind install (for viewing callgrind/cachegrind files produced by
+	# xdebug profiler)
+	if [ ! -d /srv/www/default/webgrind ]
+	then
+		echo -e "\nDownloading webgrind, see https://github.com/jokkedk/webgrind"
+		git clone git://github.com/jokkedk/webgrind.git /srv/www/default/webgrind
+	else
+		echo -e "\nUpdating webgrind..."
+		cd /srv/www/default/webgrind
+		git pull --rebase origin master
+	fi
+
 	# Install and configure the latest stable version of WordPress
 	if [ ! -d /srv/www/wordpress-default ]
 	then
-		printf "Downloading WordPress.....http://wordpress.org\n"
+		echo "Downloading WordPress Stable, see http://wordpress.org/"
 		cd /srv/www/
 		curl -O http://wordpress.org/latest.tar.gz
 		tar -xvf latest.tar.gz
 		mv wordpress wordpress-default
 		rm latest.tar.gz
 		cd /srv/www/wordpress-default
-		printf "Configuring WordPress...\n"
+		echo "Configuring WordPress Stable..."
 		wp core config --dbname=wordpress_default --dbuser=wp --dbpass=wp --quiet --extra-php <<PHP
-define( "WP_DEBUG", true );
+define( 'WP_DEBUG', true );
 PHP
 		wp core install --url=local.wordpress.dev --quiet --title="Local WordPress Dev" --admin_name=admin --admin_email="admin@local.dev" --admin_password="password"
 	else
-		printf "Skip WordPress installation, already available\n"
+		echo "Updating WordPress Stable..."
+		cd /srv/www/wordpress-default
+		wp core upgrade
 	fi
 
-	# Checkout, install and configure WordPress trunk
+	# Checkout, install and configure WordPress trunk via core.svn
 	if [ ! -d /srv/www/wordpress-trunk ]
 	then
-		printf "Checking out WordPress trunk....http://core.svn.wordpress.org/trunk\n"
+		echo "Checking out WordPress trunk from core.svn, see http://core.svn.wordpress.org/trunk"
 		svn checkout http://core.svn.wordpress.org/trunk/ /srv/www/wordpress-trunk
 		cd /srv/www/wordpress-trunk
-		printf "Configuring WordPress trunk...\n"
+		echo "Configuring WordPress trunk..."
 		wp core config --dbname=wordpress_trunk --dbuser=wp --dbpass=wp --quiet --extra-php <<PHP
-define( "WP_DEBUG", true );
+define( 'WP_DEBUG', true );
 PHP
 		wp core install --url=local.wordpress-trunk.dev --quiet --title="Local WordPress Trunk Dev" --admin_name=admin --admin_email="admin@local.dev" --admin_password="password"
 	else
-		printf "Updating WordPress trunk...\n"
+		echo "Updating WordPress trunk..."
 		cd /srv/www/wordpress-trunk
 		svn up --ignore-externals
 	fi
 
-	# Checkout and configure the WordPress unit tests
-	if [ ! -f /home/vagrant/flags/disable_wp_tests ]
+	# Checkout, install and configure WordPress trunk via develop.svn
+	if [ ! -d /srv/www/wordpress-develop ]
 	then
-		if [ ! -d /srv/www/wordpress-unit-tests ]
-		then
-			printf "Downloading WordPress Unit Tests.....https://unit-tests.svn.wordpress.org\n"
-			# Must be in a WP directory to run wp
-			cd /srv/www/wordpress-trunk
-			wp core init-tests /srv/www/wordpress-unit-tests --dbname=wordpress_unit_tests --dbuser=wp --dbpass=wp
-		else
-			printf "Updating WordPress unit tests...\n"	
-			cd /srv/www/wordpress-unit-tests
-			svn up --ignore-externals
-		fi
+		echo "Checking out WordPress trunk from develop.svn, see http://develop.svn.wordpress.org/trunk"
+		svn checkout http://develop.svn.wordpress.org/trunk/ /srv/www/wordpress-develop
+		cd /srv/www/wordpress-develop/src/
+		echo "Configuring WordPress develop..."
+		wp core config --dbname=wordpress_develop --dbuser=wp --dbpass=wp --quiet --extra-php <<PHP
+define( 'WP_DEBUG', true );
+PHP
+		wp core install --url=src.wordpress-develop.dev --quiet --title="WordPress Develop" --admin_name=admin --admin_email="admin@local.dev" --admin_password="password"
+		cp /srv/config/wordpress-config/wp-tests-config.php /srv/www/wordpress-develop/
+	else
+		echo "Updating WordPress trunk..."
+		cd /srv/www/wordpress-develop/
+		svn up
 	fi
 
-	# Download phpMyAdmin 4.0.3
+	if [ ! -d /srv/www/wordpress-develop/build ]
+	then
+		echo "Initializing grunt in WordPress develop..."
+		cd /srv/www/wordpress-develop/
+		npm install
+		grunt
+	fi
+
+	# Download phpMyAdmin 4.0.5
 	if [ ! -d /srv/www/default/database-admin ]
 	then
-		printf "Downloading phpMyAdmin 4.0.3....\n"
+		echo "Downloading phpMyAdmin 4.0.5..."
 		cd /srv/www/default
-		wget -q -O phpmyadmin.tar.gz 'http://sourceforge.net/projects/phpmyadmin/files/phpMyAdmin/4.0.3/phpMyAdmin-4.0.3-english.tar.gz/download#!md5!07dc6ed4d65488661d2581de8d325493'
+		wget -q -O phpmyadmin.tar.gz 'http://sourceforge.net/projects/phpmyadmin/files/phpMyAdmin/4.0.5/phpMyAdmin-4.0.5-english.tar.gz/download'
 		tar -xf phpmyadmin.tar.gz
-		mv phpMyAdmin-4.0.3-english database-admin
+		mv phpMyAdmin-4.0.5-english database-admin
 		rm phpmyadmin.tar.gz
 	else
-		printf "PHPMyAdmin 4.0.3 already installed.\n"
+		echo "PHPMyAdmin already installed."
 	fi
 else
-	printf "\nNo network available, skipping network installations"
+	echo -e "\nNo network available, skipping network installations"
 fi
-# Add any custom domains to the virtual machine's hosts file so that it
-# is self aware. Enter domains space delimited as shown with the default.
 
-echo "# Hosts for varying-vagrant-vagrants:" > /srv/config/host-hosts
+echo "# Hosts for varying-vagrant-vagrants (put in host machine's hosts file):" > /srv/config/host-hosts
 echo "# Make sure the following exist in your host machine's hosts file"
 echo "# Full list will be located in config/host-hosts"
-wp_domains='local.wordpress.dev local.wordpress-trunk.dev'
-if ! grep -q "$wp_domains" /etc/hosts
+
+# Add any custom domains to the virtual machine's hosts file so that it
+# is self aware. Enter domains space delimited as shown with the default.
+DOMAINS='local.wordpress.dev 
+         local.wordpress-trunk.dev
+         src.wordpress-develop.dev
+         build.wordpress-develop.dev'
+
+if ! grep -q "$DOMAINS" /etc/hosts
 then
-	echo "127.0.0.1 $wp_domains" >> /etc/hosts
+	DOMAINS=$(echo $DOMAINS)
+	echo "127.0.0.1 $DOMAINS" >> /etc/hosts
 fi
-echo "$vagrant_ip $wp_domains" | tee -a /srv/config/host-hosts
+echo "$vvv_ip $(echo $DOMAINS)" | tee -a /srv/config/host-hosts
 
 # Look for additional domains defined in the sites
 for site_domains_file in $(find /srv/www -name 'vvv-domains')
@@ -411,17 +511,17 @@ do
 		then
 			echo "127.0.0.1 $site_domain_line" >> /etc/hosts
 		fi
-		echo "$vagrant_ip $site_domain_line" | tee -a /srv/config/host-hosts
+		echo "$vvv_ip $site_domain_line" | tee -a /srv/config/host-hosts
 	done
 done
 
 end_seconds=`date +%s`
-echo -----------------------------
-echo Provisioning complete in `expr $end_seconds - $start_seconds` seconds
+echo "-----------------------------"
+echo "Provisioning complete in `expr $end_seconds - $start_seconds` seconds"
 if [[ $ping_result == *bytes?from* ]]
 then
-	echo External network connection established, packages up to date.
+	echo "External network connection established, packages up to date."
 else
-	echo No external network available. Package installation and maintenance skipped.
+	echo "No external network available. Package installation and maintenance skipped."
 fi
-echo For further setup instructions, visit http://$vagrant_ip
+echo "For further setup instructions, visit http://$vvv_ip"