+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Varying Vagrant Vagrants Changelog
 
 ## 0.9-working
+* Use `precise32` for the Vagrant box name for better cross project box caching.
+    * **Note:** This will probably cause a new Vagrant box to download. Use `vagrant box remove std-precise32` after a `vagrant destroy` to remove the old one and start with this.
+* **Possible Breaking:** Remove flags system entirely.
+	* **Note:** The only remaining flag was for the WordPress unit tests. The functionality of this flag can be duplicated by creating the empty directory `{vvv_dir}/www/wordpress-unit-tests`.
+* **Possible Breaking:** Change VM hostname to `vvv`
+    * **Note:** If you had anything setup to rely on the hostname of precise32-dev, this may break.
+* Update stable version of WordPress automatically on provision
+* Provide default certs for SSL in Nginx
+* Update to phpMyAdmin 4.0.5
 
 ## 0.8
 * Enable SSH agent forwarding